Add support for the Keep a Changelog format
This format checker is a proof-of-concept for the format and uses the
example change log from the website as its test. There is likely some
work we could do to make the format checker more robust and, for
example, integrate it with the Intridea format checker for the line
items.